Added option to output probability map from CLI

+++ lungtumormask/mask.py
@@ -27,7 +27,7 @@ def mask(image_path, save_path, lung_filter, threshold, radius, batch_size):
     right = model(preprocess_dump['right_lung']).squeeze(0).squeeze(0).detach().numpy()
 
     print("Post-processing image...")
-    inferred = post_process(left, right, preprocess_dump, lung_filter, threshold, radius).astype("uint8")
+    inferred = post_process(left, right, preprocess_dump, lung_filter, threshold, radius)
 
     print(f"Storing segmentation at {save_path}")
     nimage = nibabel.Nifti1Image(inferred, preprocess_dump['org_affine'])
